requirements.txt
pillow
numpy==1.16.0
opencv-python-headless==4.2.0.32
streamlit
tensorflow